iox-#33 created platform library and refactored sem_open to be more fitting to the posix implementation otherwise we get template deduction problems between platforms like windows and linux
